agent/pyproject.toml
[project]
name = "multi-agent-assistant"
version = "1.0.0"
description = "Multi-agent system with 100% open source components"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.0,<2.0.0",
    "langgraph>=1.0.5,<2.0.0",
    "langchain-google-genai>=4.1.3",
    "langchain-openai>=1.1.1",
    "fastapi>=0.115.5,<1.0.0",
    "uvicorn[standard]>=0.29.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "copilotkit>=0.1.74",
    "ag-ui-langgraph>=0.0.22",
    "arize-phoenix>=12.28.1",
    "openinference-instrumentation-langchain>=0.1.58",
]